+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/ApiVersionsServiceImpl.java
@@ -12,13 +12,13 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 
-import org.apache.kafka.common.message.ApiVersionsRequestData;
 import org.apache.kafka.common.message.ApiVersionsResponseData;
 import org.apache.kafka.common.message.ApiVersionsResponseData.ApiVersion;
 import org.apache.kafka.common.message.ApiVersionsResponseData.ApiVersionCollection;
 import org.apache.kafka.common.message.RequestHeaderData;
 import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.Errors;
+import org.apache.kafka.common.requests.ApiVersionsRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -31,6 +31,7 @@ public class ApiVersionsServiceImpl {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ApiVersionsServiceImpl.class);
     private ApiVersions apiVersions = null;
+    private ApiVersionsRequest apiVersionsRequest = null;
 
     public void updateVersions(String channel, ApiVersionsResponseData upstreamApiVersions) {
         var upstream = upstreamApiVersions.duplicate().apiKeys();
@@ -105,9 +106,9 @@ public class ApiVersionsServiceImpl {
 
         // KIP-511 when the client receives an unsupported version for the ApiVersionResponse, it fails back to version 0
         // Use the same algorithm as https://github.com/apache/kafka/blob/159d25a7df25975694e2e0eb18a8feb125f7c39e/clients/src/main/java/org/apache/kafka/clients/NetworkClient.java#L957-L977
-        var data = new ApiVersionsRequestData();
-        var header = new RequestHeaderData().setRequestApiVersion(data.highestSupportedVersion());
-        return context.<ApiVersionsResponseData> sendRequest(header, data)
+        var request = getApiVersionsRequest();
+        var header = new RequestHeaderData().setRequestApiVersion(request.version());
+        return context.<ApiVersionsResponseData> sendRequest(header, request.data())
                 .thenCompose(response -> {
                     if (response.errorCode() != Errors.NONE.code()) {
                         if (header.requestApiVersion() == 0 || response.errorCode() != Errors.UNSUPPORTED_VERSION.code()) {
@@ -122,7 +123,7 @@ public class ApiVersionsServiceImpl {
                                 maxApiVersion = apiVersion.maxVersion();
                             }
                         }
-                        return context.sendRequest(header.setRequestApiVersion(maxApiVersion), data);
+                        return context.sendRequest(header.setRequestApiVersion(maxApiVersion), request.data());
                     }
                     return CompletableFuture.completedStage(response);
                 })
@@ -130,6 +131,25 @@ public class ApiVersionsServiceImpl {
                     updateVersions(context.channelDescriptor(), response);
                     return apiVersions;
                 });
+    }
+
+    void setApiVersionsRequest(ApiVersionsRequest request) {
+        if (request.hasUnsupportedRequestVersion()) {
+            throw Errors.UNSUPPORTED_VERSION.exception();
+        }
+        if (!request.isValid()) {
+            throw Errors.INVALID_REQUEST.exception();
+        }
+        this.apiVersionsRequest = request;
+    }
+
+    ApiVersionsRequest getApiVersionsRequest() {
+        if (apiVersionsRequest == null) {
+            ApiVersionsRequest.Builder builder = new ApiVersionsRequest.Builder();
+            short version = builder.latestAllowedVersion();
+            apiVersionsRequest = builder.build(version);
+        }
+        return apiVersionsRequest;
     }
 
 }