+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/internal/ApiVersionsServiceImplTest.java
@@ -9,17 +9,23 @@ package io.kroxylicious.proxy.internal;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
+import org.apache.kafka.common.errors.InvalidRequestException;
+import org.apache.kafka.common.errors.UnsupportedVersionException;
+import org.apache.kafka.common.message.ApiVersionsRequestData;
 import org.apache.kafka.common.message.ApiVersionsResponseData;
 import org.apache.kafka.common.message.RequestHeaderData;
 import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.Errors;
+import org.apache.kafka.common.requests.ApiVersionsRequest;
 import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 
 import io.kroxylicious.proxy.ApiVersionsService;
 import io.kroxylicious.proxy.filter.FilterContext;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.argThat;
 
@@ -76,6 +82,24 @@ class ApiVersionsServiceImplTest {
         assertThat(range.upstream().maxVersion()).isEqualTo((short) (ApiKeys.METADATA.latestVersion() - 1));
         assertThat(range.intersected().minVersion()).isEqualTo(ApiKeys.METADATA.oldestVersion());
         assertThat(range.intersected().maxVersion()).isEqualTo((short) (ApiKeys.METADATA.latestVersion() - 1));
+    }
+
+    @Test
+    void testGetVersionRanges_ApiVersionsRequestIsValid() {
+        ApiVersionsServiceImpl apiVersionsService = new ApiVersionsServiceImpl();
+        FilterContext filterContext = Mockito.mock(FilterContext.class);
+        ApiVersionsResponseData upstreamApiVersions = createApiVersionsWith(ApiKeys.METADATA.id, ApiKeys.METADATA.oldestVersion(),
+                ApiKeys.METADATA.latestVersion());
+        Mockito.when(filterContext.sendRequest(any(RequestHeaderData.class), any()))
+                .thenReturn(CompletableFuture.completedFuture(upstreamApiVersions));
+        apiVersionsService.getApiVersionRanges(ApiKeys.METADATA, filterContext).toCompletableFuture()
+                .getNow(Optional.empty());
+        ArgumentCaptor<RequestHeaderData> requestHeader = ArgumentCaptor.forClass(RequestHeaderData.class);
+        ArgumentCaptor<ApiVersionsRequestData> requestData = ArgumentCaptor.forClass(ApiVersionsRequestData.class);
+        Mockito.verify(filterContext).sendRequest(requestHeader.capture(), requestData.capture());
+        ApiVersionsRequest apiVersionsRequest = new ApiVersionsRequest(requestData.getValue(), requestHeader.getValue().requestApiVersion());
+        assertThat(apiVersionsRequest.hasUnsupportedRequestVersion()).isFalse();
+        assertThat(apiVersionsRequest.isValid()).isTrue();
     }
 
     @Test
@@ -147,6 +171,47 @@ class ApiVersionsServiceImplTest {
         upstreamApiVersions.apiKeys().add(new ApiVersionsResponseData.ApiVersion().setApiKey(api).setMinVersion(minVersion).setMaxVersion(
                 maxVersion));
         return upstreamApiVersions;
+    }
+
+    @Test
+    void testSetApiVersionsRequest_SetAndGetApiVersionsRequest() {
+        ApiVersionsServiceImpl apiVersionsService = new ApiVersionsServiceImpl();
+        ApiVersionsRequest apiVersionsRequest = new ApiVersionsRequest(new ApiVersionsRequestData()
+                .setClientSoftwareName("my-test-client")
+                .setClientSoftwareVersion("1.0.0"), (short) 3);
+        apiVersionsService.setApiVersionsRequest(apiVersionsRequest);
+        assertThat(apiVersionsService.getApiVersionsRequest()).isSameAs(apiVersionsRequest);
+    }
+
+    @Test
+    void testGetApiVersionsRequest_InvalidRequest() {
+        ApiVersionsServiceImpl apiVersionsService = new ApiVersionsServiceImpl();
+
+        ApiVersionsRequest unsupportedApiVersion = new ApiVersionsRequest(new ApiVersionsRequestData()
+                .setClientSoftwareName("my-test-client")
+                .setClientSoftwareVersion("1.0.0"), (short) 0, (short) (ApiKeys.API_VERSIONS.latestVersion() + 1));
+        assertThatThrownBy(() -> apiVersionsService.setApiVersionsRequest(unsupportedApiVersion))
+                .isInstanceOf(UnsupportedVersionException.class);
+
+        ApiVersionsRequest invalidClientSoftwareName = new ApiVersionsRequest(new ApiVersionsRequestData()
+                .setClientSoftwareName("$my-test-client$")
+                .setClientSoftwareVersion("1.0.0"), (short) 3);
+        assertThatThrownBy(() -> apiVersionsService.setApiVersionsRequest(invalidClientSoftwareName))
+                .isInstanceOf(InvalidRequestException.class);
+
+        ApiVersionsRequest invalidClientSoftwareVersion = new ApiVersionsRequest(new ApiVersionsRequestData()
+                .setClientSoftwareName("my-test-client")
+                .setClientSoftwareVersion("$1.0.0$"), (short) 3);
+        assertThatThrownBy(() -> apiVersionsService.setApiVersionsRequest(invalidClientSoftwareVersion))
+                .isInstanceOf(InvalidRequestException.class);
+    }
+
+    @Test
+    void testGetApiVersionsRequest_DefaultApiVersionsRequest() {
+        ApiVersionsServiceImpl apiVersionsService = new ApiVersionsServiceImpl();
+        ApiVersionsRequest request = apiVersionsService.getApiVersionsRequest();
+        assertThat(request.version()).isEqualTo(ApiKeys.API_VERSIONS.latestVersion());
+        assertThat(request.data().clientSoftwareName()).isEqualTo("apache-kafka-java");
     }
 
 }